+++ components/anamnese/ExportPDFButton/index.ts
@@ -1,0 +1,2 @@
+export { ExportPDFButton } from './ExportPDFButton'
+export { useExportPdf } from './useExportPdf'